+++ service/cmd/main.go
@@ -1,51 +1,22 @@
 package main
 
 import (
-	"context"
 	"github.com/mmcdole/gofeed"
 	"github.com/terratensor/feed-parser/service/internal/entities/feed"
+	"github.com/terratensor/feed-parser/service/internal/parser"
 	"github.com/terratensor/feed-parser/service/internal/workerpool"
 	"log"
-	"math/rand"
 	"sync"
 	"time"
 )
 
-const (
-	envLocal = "local"
-	envDev   = "dev"
-	envProd  = "prod"
-)
-
-type Link struct {
-	Url        string
-	Lang       string
-	ResourceID int
-}
-
 func main() {
-
-	// можно протащить logger через value контекста, и далее вызывать его не передавая
-	// в качестве параметра функции или поля структуры, например, используя logger := ctx.Value("logger").(slog.Logger)
-	//ctx, cancel := signal.NotifyContext(context.WithValue(context.Background(), "logger", *logger), os.Interrupt)
-	//defer cancel()
-	//
-	//var storage feed.StorageInterface
-	//
-	//manticoreClient, err := manticore.New("feed")
-	//if err != nil {
-	//	log.Fatalf("failed to initialize manticore client, %v", err)
-	//	os.Exit(1)
-	//}
-	//
-	//storage = manticoreClient
-	wg := &sync.WaitGroup{}
 
 	fp := gofeed.NewParser()
 	//fp.UserAgent = "PostmanRuntime/7.36.3"
 	ch := make(chan feed.Entry, 20)
 
-	urls := []Link{
+	urls := []parser.Link{
 		{Url: "http://kremlin.ru/events/all/feed/", Lang: "ru", ResourceID: 1},
 		{Url: "http://en.kremlin.ru/events/all/feed", Lang: "en", ResourceID: 1},
 		{Url: "https://mid.ru/ru/rss.php", Lang: "ru", ResourceID: 2},
@@ -61,9 +32,11 @@ func main() {
 		//{Url: "https://tass.com/rss/v2.xml", Lang: "en", ResourceID: 4},
 	}
 
+	wg := &sync.WaitGroup{}
 	for _, url := range urls {
 		wg.Add(1)
-		go runParser(url, ch, fp, wg)
+		p := parser.NewParser(url, 60*time.Second, 150*time.Second)
+		go p.Run(ch, fp, wg)
 	}
 
 	var allTask []*workerpool.Task
@@ -89,62 +62,4 @@ func main() {
 
 	wg.Wait()
 	log.Println("finished, all workers successfully stopped.")
-}
-
-func runParser(url Link, ch chan feed.Entry, fp *gofeed.Parser, wg *sync.WaitGroup) {
-	defer wg.Done()
-
-	for {
-
-		n := 60 + rand.Intn(150)
-		d := time.Duration(n)
-		time.Sleep(d * time.Second)
-
-		log.Printf("started parser for given url: %v", url.Url)
-		gf, err := fp.ParseURL(url.Url)
-		if err != nil {
-			log.Printf("ERROR: %v, %v", err, url.Url)
-			continue
-		}
-		log.Printf("fetched the contents of a given url %v", url.Url)
-		entries := makeEntries(gf.Items, url)
-
-		select {
-		case <-context.Background().Done():
-			break
-		default:
-		}
-
-		for _, entry := range entries {
-			ch <- entry
-		}
-	}
-}
-
-func makeEntries(items []*gofeed.Item, url Link) []feed.Entry {
-	var entries []feed.Entry
-
-	for _, item := range items {
-
-		e := &feed.Entry{
-			Language:   url.Lang,
-			Title:      item.Title,
-			Url:        item.Link,
-			Updated:    item.UpdatedParsed,
-			Published:  item.PublishedParsed,
-			Summary:    item.Description,
-			Content:    item.Content,
-			ResourceID: url.ResourceID,
-		}
-
-		entries = append(entries, *e)
-	}
-
-	return entries
-}
-
-func waiting(i int) {
-	n := 1 * rand.Intn(i)
-	d := time.Duration(n)
-	time.Sleep(d * time.Second)
 }
